Use direct search method to get section recents.
Changes in the Plex server mean the currently-released version of the plexapi library returns a list of shows when queried for recently added items, rather than episodes. An unreleased version of the api contains a fix, but plexlib now uses an own help method that searches directly using appropriate keywords.